Updating the generator to be more testable and configurable
