+++ generator/loadmeta.go
@@ -18,8 +18,8 @@ package generator
 
 import (
 	"bytes"
-	"encoding/json"
-	"flag"
+
+
 	"fmt"
 	"go/format"
 	"io/ioutil"
@@ -28,239 +28,24 @@ import (
 	"sort"
 	"strings"
 	"text/template"
-
 	"golang.org/x/tools/imports"
+	"github.com/spf13/cobra"
 )
 
-const license = `/**
- * Copyright 2016-2024 IBM Corp.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with 
- * the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed 
- * on an "AS IS" BASIS,WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and limitations under the License.
- */
- `
-const codegenWarning = `// AUTOMATICALLY GENERATED CODE - DO NOT MODIFY`
-const SoftLayerMetadataAPIURL = "https://api.softlayer.com/metadata/v3.1"
+var rootCmd = &cobra.Command{
+	Use: "tools",
+	Short: "Generates the services and datatype definitions from SoftLayer API Metadata",
 
-type Type struct {
-	Name         string              `json:"name"`
-	Base         string              `json:"base"`
-	TypeDoc      string              `json:"typeDoc"`
-	Properties   map[string]Property `json:"properties"`
-	ServiceDoc   string              `json:"serviceDoc"`
-	Methods      map[string]Method   `json:"methods"`
-	NoService    bool                `json:"noservice"`
-	Overview     string              `json:"docOverview"`
-	Deprecated   bool                `json:"deprecated"`
-	ServiceGroup string
 }
 
-type Property struct {
-	Name       string `json:"name"`
-	Type       string `json:"type"`
-	TypeArray  bool   `json:"typeArray"`
-	Form       string `json:"form"`
-	Doc        string `json:"doc"`
-	Overview   string `json:"docOverview"`
-	Deprecated bool   `json:"deprecated"`
+func Execute() error {
+	return rootCmd.Execute()
 }
 
-type Method struct {
-	Name       string      `json:"name"`
-	Type       string      `json:"type"`
-	TypeArray  bool        `json:"typeArray"`
-	Doc        string      `json:"doc"`
-	Static     bool        `json:"static"`
-	NoAuth     bool        `json:"noauth"`
-	Limitable  bool        `json:"limitable"`
-	Filterable bool        `json:"filterable"`
-	Maskable   bool        `json:"maskable"`
-	Parameters []Parameter `json:"parameters"`
-	Overview   string      `json:"docOverview"`
-	Deprecated bool        `json:"deprecated"`
+func init() {
+	// rootCmd.AddCommand(generateCmd)
+	// rootCmd.AddCommand(versionCmd)
 }
-
-type Parameter struct {
-	Name         string      `json:"name"`
-	Type         string      `json:"type"`
-	TypeArray    bool        `json:"typeArray"`
-	Doc          string      `json:"doc"`
-	DefaultValue interface{} `json:"defaultValue"`
-	ParentMethod string
-}
-
-// Define custom template functions
-var fMap = template.FuncMap{
-	"convertType":     ConvertType,         // Converts SoftLayer types to Go types
-	"removePrefix":    RemovePrefix,        // Remove 'SoftLayer_' prefix. if it exists
-	"removeReserved":  RemoveReservedWords, // Substitute language-reserved identifiers
-	"titleCase":       strings.Title,       // TitleCase the argument
-	"desnake":         Desnake,             // Remove '_' from Snake_Case
-	"goDoc":           GoDoc,               // Format a go doc string
-	"tags":            Tags,                // Remove omitempty tags if required
-	"phraseMethodArg": phraseMethodArg,     // Get proper phrase for method argument
-	"getTypePrefix":   getTypePrefix,       // resturns [], * or nothing.
-	"deprecatedDoc":   deprecatedDoc,       // Marks things as deprecated if needed.
-	"nilParam":        NilParam,            // For unit testing fakes
-}
-
-var datatype = fmt.Sprintf(`%s
-
-%s
-
-package datatypes
-
-{{range .}}{{.TypeDoc|goDoc}}
-type {{.Name|removePrefix}} struct {
-	{{.Base|removePrefix}}
-
-{{$base := .Name}}{{range .Properties}}{{.Doc|goDoc}}{{deprecatedDoc .Deprecated}}`+
-	`{{$thisType := convertType .Type "datatypes" $base .Name}}
-	{{.Name|titleCase}} {{getTypePrefix .TypeArray $thisType}}{{$thisType}}`+
-	"`json:\"{{.Name|tags}}\" xmlrpc:\"{{.Name|tags}}\"`"+`
-
-{{end}}
-}
-
-{{end}}
-`, license, codegenWarning)
-
-var services = fmt.Sprintf(`%s
-
-%s
-
-package services
-
-import (
-	"fmt"
-	"strings"
-)
-
-{{range .}}{{$base := .Name|removePrefix}}{{.TypeDoc|goDoc}}{{deprecatedDoc .Deprecated}}
-	type {{$base}} struct {
-		Session session.SLSession
-		Options sl.Options
-	}
-
-	// Get{{$base | desnake}}Service returns an instance of the {{$base}} SoftLayer service
-	func Get{{$base | desnake}}Service(sess session.SLSession) {{$base}} {
-		return {{$base}}{Session: sess}
-	}
-
-	func (r {{$base}}) Id(id int) {{$base}} {
-		r.Options.Id = &id
-		return r
-	}
-
-	func (r {{$base}}) Mask(mask string) {{$base}} {
-		if !strings.HasPrefix(mask, "mask[") && (strings.Contains(mask, "[") || strings.Contains(mask, ",")) {
-			mask = fmt.Sprintf("mask[%%s]", mask)
-		}
-
-		r.Options.Mask = mask
-		return r
-	}
-
-	func (r {{$base}}) Filter(filter string) {{$base}} {
-		r.Options.Filter = filter
-		return r
-	}
-
-	func (r {{$base}}) Limit(limit int) {{$base}} {
-		r.Options.Limit = &limit
-		return r
-	}
-
-	func (r {{$base}}) Offset(offset int) {{$base}} {
-		r.Options.Offset = &offset
-		return r
-	}
-
-	{{$rawBase := .Name}}{{range .Methods}}{{$methodName := .Name}}{{.Doc|goDoc}}{{deprecatedDoc .Deprecated}}
-	func (r {{$base}}) {{.Name|titleCase}}({{range .Parameters}}{{phraseMethodArg $methodName .Name .TypeArray .Type}}{{end}}) ({{if .Type|ne "void"}}resp {{if .TypeArray}}[]{{end}}{{convertType .Type "services"}}, {{end}}err error) {
-		{{if .Type|eq "void"}}var resp datatypes.Void
-		{{end}}{{if or (eq .Name "placeOrder") (eq .Name "verifyOrder")}}err = datatypes.SetComplexType(orderData)
-		if err != nil {
-			return
-		}
-		{{end}}{{if len .Parameters | lt 0}}params := []interface{}{
-			{{range .Parameters}}{{.Name|removeReserved}},
-			{{end}}
-		}
-		{{end}}err = r.Session.DoRequest("{{$rawBase}}", "{{.Name}}", {{if len .Parameters | lt 0}}params{{else}}nil{{end}}, &r.Options, &resp)
-	return
-	}
-	{{end}}
-
-{{end}}
-`, license, codegenWarning)
-
-var testTemplate = `// AUTO GENERATED by tools/loadmeta.go
-package services_test
-
-import (
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-	"github.com/softlayer/softlayer-go/services"
-	"github.com/softlayer/softlayer-go/session/sessionfakes"
-)
-
-var _ = Describe("{{(index . 0 ).ServiceGroup}} Tests", func() {
-	var slsession *sessionfakes.FakeSLSession
-	BeforeEach(func() {
-		slsession = &sessionfakes.FakeSLSession{}
-	})
-{{range .}}
-	Context("Testing {{.Name}} service", func() {
-		var sl_service services.{{.Name | removePrefix}}
-		BeforeEach(func() {
-			sl_service = services.Get{{.Name | removePrefix | desnake}}Service(slsession)
-		})
-{{- $serviceName := .Name }}
-		Context("{{$serviceName}} Set Options", func() {
-			It("Set Options properly", func() {
-				t_id := 1234
-				t_filter := "{'testFilter':{'test'}}"
-				t_limit := 100
-				t_offset := 5
-				sl_service = sl_service.Id(t_id).Filter(t_filter).Offset(t_offset).Limit(t_limit)
-				Expect(sl_service.Options.Id).To(HaveValue(Equal(t_id)))
-				Expect(sl_service.Options.Filter).To(HaveValue(Equal(t_filter)))
-				Expect(sl_service.Options.Limit).To(HaveValue(Equal(t_limit)))
-				Expect(sl_service.Options.Offset).To(HaveValue(Equal(t_offset)))
-			})
-		})
-		Context("{{$serviceName}} Set Mask", func() {
-			It("Set Options properly", func() {
-				t_mask1 := "mask[test,test2]"
-				sl_service = sl_service.Mask(t_mask1)
-				Expect(sl_service.Options.Mask).To(HaveValue(Equal(t_mask1)))
-				// Mask("test,test2") should set the mask to be "mask[test,test2]" aka t_mask1
-				sl_service = sl_service.Mask("test,test2")
-				Expect(sl_service.Options.Mask).To(HaveValue(Equal(t_mask1)))
-			})
-		})
-{{- range .Methods }}
-		Context("{{$serviceName}}::{{.Name}}", func() {
-			It("API Call Test", func() {
-{{- if .Type|ne "void"}}
-				_, err := sl_service.{{.Name | titleCase}}({{.Parameters | nilParam}})
-{{- else }}
-				err := sl_service.{{.Name | titleCase}}({{.Parameters | nilParam}})
-{{- end }}
-				Expect(err).To(Succeed())
-				Expect(slsession.DoRequestCallCount()).To(Equal(1))
-			})
-		})
-{{- end }}
-	})
-{{ end }}
-})
-`
 
 func NilParam(params []Parameter) string {
 	rString := ""
@@ -278,99 +63,6 @@ func NilParam(params []Parameter) string {
 	}
 	return rString
 }
-
-func GenerateAPI() {
-	var meta map[string]Type
-
-	flagset := flag.NewFlagSet(os.Args[1], flag.ExitOnError)
-	outputPath := flag.String("o", ".", "the root of the go project to be refreshed")
-	flagset.Parse(os.Args[2:])
-
-	jsonResp, code, err := makeHttpRequest(SoftLayerMetadataAPIURL, "GET", new(bytes.Buffer))
-
-	if err != nil {
-		fmt.Printf("Error retrieving metadata API: %s", err)
-		os.Exit(1)
-	}
-
-	if code != 200 {
-		fmt.Printf("Unexpected HTTP status code received while retrieving metadata API: %d", code)
-		os.Exit(1)
-	}
-
-	err = json.Unmarshal(jsonResp, &meta)
-	if err != nil {
-		fmt.Printf("Error unmarshaling json response: %s", err)
-		os.Exit(1)
-	}
-
-	// Build an array of Types, sorted by name
-	// This will ensure consistency in the order that code is later emitted
-	keys := getSortedKeys(meta)
-
-	sortedTypes := make([]Type, 0, len(keys))
-	sortedServices := make([]Type, 0, len(keys))
-
-	for _, name := range keys {
-		t := meta[name]
-		sortedTypes = append(sortedTypes, t)
-		addComplexType(&t)
-		fixDatatype(&t, meta)
-
-		// Not every datatype is also a service
-		if !t.NoService {
-			createGetters(&t)
-			sortedServices = append(sortedServices, t)
-		}
-	}
-
-	// Services can be subclasses of other services. Copy methods from each service's 'Base' entity to
-	// the child service, only if a same-named method does not already exist (i.e., overridden by the
-	// child service)
-	for i, service := range sortedServices {
-		sortedServices[i].Methods = getBaseMethods(service, meta)
-		fixReturnType(&sortedServices[i])
-	}
-
-	err = writePackage(*outputPath, "datatypes", sortedTypes, datatype)
-	if err != nil {
-		fmt.Printf("Error writing to file: %s", err)
-	}
-
-	err = writePackage(*outputPath, "services", sortedServices, services)
-	if err != nil {
-		fmt.Printf("Error writing to file: %s", err)
-	}
-
-}
-
-func generateTestFiles(meta []Type, groupName string) {
-	// These are technically a service... but dont have any methods to test...
-	if groupName == "exception" {
-		return
-	}
-	path := fmt.Sprintf("./services/%s_test.go", groupName)
-	template := template.New(path)
-	template.Funcs(fMap)
-	_, err := template.Parse(testTemplate)
-	if err != nil {
-		fmt.Printf("Error Parsing Template: %s\n", err)
-		return
-	}
-	outfile, err := os.Create(path)
-	if err != nil {
-		fmt.Printf("Error Creating File %s: %s\n", path, err)
-		return
-	}
-	defer outfile.Close()
-	err = template.Execute(outfile, meta)
-	if err != nil {
-		fmt.Printf("Error Executing Template: %s\n", err)
-		return
-	}
-}
-
-// Exported template functions
 
 func RemovePrefix(args ...interface{}) string {
 	s := args[0].(string)
@@ -659,7 +351,7 @@ func writePackage(base string, pkg string, meta []Type, ts string) error {
 // Executes a template against the metadata structure, and generates a go source file with the result
 func writeGoFile(base string, pkg string, name string, meta []Type, ts string) error {
 	filename := base + "/" + pkg + "/" + strings.ToLower(name) + ".go"
-
+	fmt.Printf("Creating %s\n", filename)
 	// Generate the source
 	var buf bytes.Buffer
 	t := template.New(pkg).Funcs(fMap)
@@ -683,38 +375,65 @@ func writeGoFile(base string, pkg string, name string, meta []Type, ts string) e
 
 	f, err := os.Create(filename)
 	if err != nil {
-		return fmt.Errorf("Error creating file: %s", err)
+		return fmt.Errorf("Error creating %s: %s", filename, err)
 	}
 	defer f.Close()
 	fmt.Fprintf(f, "%s", pretty)
 	if pkg == "services" {
-		generateTestFiles(meta, strings.ToLower(name))
+		generateTestFiles(base, meta, strings.ToLower(name))
 	}
 
 	return nil
 }
 
-func makeHttpRequest(url string, requestType string, requestBody *bytes.Buffer) ([]byte, int, error) {
+func generateTestFiles(base string, meta []Type, groupName string) {
+	// These are technically a service... but dont have any methods to test...
+	if groupName == "exception" {
+		return
+	}
+	path := fmt.Sprintf("%s/services/%s_test.go", base, groupName)
+	fmt.Printf("Creating %s\n", path)
+	template := template.New(path)
+	template.Funcs(fMap)
+	_, err := template.Parse(testTemplate)
+	if err != nil {
+		fmt.Printf("Error Parsing Template: %s\n", err)
+		return
+	}
+	outfile, err := os.Create(path)
+	if err != nil {
+		fmt.Printf("Error Creating File %s: %s\n", path, err)
+		return
+	}
+	defer outfile.Close()
+	err = template.Execute(outfile, meta)
+	if err != nil {
+		fmt.Printf("Error Executing Template: %s\n", err)
+		return
+	}
+}
+
+func GetMetaFromURL(url string) ([]byte, error) {
 	client := http.DefaultClient
 
-	req, err := http.NewRequest(requestType, url, requestBody)
+	req, err := http.NewRequest("GET", url, nil)
 	if err != nil {
-		return nil, 0, err
+		return nil, err
 	}
 
 	resp, err := client.Do(req)
 	if err != nil {
-		return nil, 520, err
+		return nil, err
 	}
 
 	defer resp.Body.Close()
 
 	responseBody, err := ioutil.ReadAll(resp.Body)
 	if err != nil {
-		return nil, resp.StatusCode, err
+		return nil, err
 	}
 
-	return responseBody, resp.StatusCode, nil
+	return responseBody, nil
 }
 
 func getTypePrefix(isArray bool, theType string) string {
