+++ tools/main.go
@@ -17,9 +17,8 @@
 package main
 
 import (
-	"fmt"
-	. "github.com/softlayer/softlayer-go/generator"
-	"os"
+
+	"github.com/softlayer/softlayer-go/generator"
 )
 
 const usage = `Usage: tools <cmd> [options]
@@ -31,19 +30,5 @@ Commands:
 	version: library version management`
 
 func main() {
-	if len(os.Args) < 2 {
-		fmt.Println(usage)
-		os.Exit(1)
-	}
-
-	switch os.Args[1] {
-	case "generate":
-		GenerateAPI()
-	case "version":
-		version()
-	default:
-		fmt.Println("Unrecognized command")
-		os.Exit(1)
-	}
-
+	generator.Execute()
 }
